incorporating review feedback
